test: PostCard の通報系画面試験を PostCard.report.test.tsx へ分割する
PostCard.test.tsx が 1000 行を超えたため(oversized-files ゲート)、通報経路(投稿・画像添付・
動画添付・最新 manifest 取得)の試験を分け、共通の表示データを PostCard.testHelpers.ts に置く。
